best_drbd_kmod: make this a classmethod
This allows us to easily call it with a forced distribution name. Centos
sucks, as it does not provide a sane /etc/os-release. But we have a
special case within a container where we know we are executed on a "RHEL
based" host OS (otherwise another container would have been started). In
this case we can just force a good enough name and run the kmod
detection.