fix linbit-reponame
This is supposed to return the name of the distribution as we have it on
packages.linbit.com. For Debian/Ubuntu this returned the name as we have
it in lbbuild (e.g., ubuntu-focal), but on p.l.c we have proper names,
at least for the Ubuntu/Debian ones (e.g., focal).